requirements.txt
django
django-autoslug
django-stdimage
django-grappelli
django_compressor
django-sortable
django-adminfiles
markdown
yammy
feedparser
purl

sorl-thumbnail
flickrapi
gdata
simplejson
httplib2
pillow
beautifulsoup4
python-memcached
pycrypto

-e git+https://github.com/praekelt/django-preferences.git#egg=django-preferences
